move testing utils under tests
