+++ tests/models_test.py
@@ -25,7 +25,7 @@ from simpeg_drivers.potential_fields import MVIInversionOptions
 from simpeg_drivers.potential_fields.magnetic_vector.driver import (
     MVIInversionDriver,
 )
-from simpeg_drivers.utils.testing import Geoh5Tester, setup_inversion_workspace
+from tests.testing_utils import Geoh5Tester, setup_inversion_workspace
 
 
 def get_mvi_params(tmp_path: Path) -> MVIInversionOptions:
