+++ tests/run_tests/driver_fem_test.py
@@ -28,8 +28,8 @@ from simpeg_drivers.electromagnetics.frequency_domain.options import (
     FDEMInversionOptions,
 )
 from simpeg_drivers.options import ActiveCellsOptions
-from simpeg_drivers.utils.testing import check_target, setup_inversion_workspace
 from simpeg_drivers.utils.utils import get_inversion_output
+from tests.testing_utils import check_target, setup_inversion_workspace
 
 
 # To test the full run and validate the inversion.
