+++ tests/run_tests/driver_ip_test.py
@@ -23,8 +23,8 @@ from simpeg_drivers.electricals.induced_polarization.three_dimensions.driver imp
     IP3DInversionDriver,
 )
 from simpeg_drivers.options import ActiveCellsOptions
-from simpeg_drivers.utils.testing import check_target, setup_inversion_workspace
 from simpeg_drivers.utils.utils import get_inversion_output
+from tests.testing_utils import check_target, setup_inversion_workspace
 
 
 # To test the full run and validate the inversion.
