+++ tests/run_tests/sensitivity_cutoff_test.py
@@ -24,7 +24,7 @@ from simpeg_drivers.depth_of_investigation.sensitivity_cutoff.options import (
 from simpeg_drivers.options import ActiveCellsOptions
 from simpeg_drivers.potential_fields import GravityInversionOptions
 from simpeg_drivers.potential_fields.gravity.driver import GravityInversionDriver
-from simpeg_drivers.utils.testing import setup_inversion_workspace
+from tests.testing_utils import setup_inversion_workspace
 
 
 def setup_inversion_results(
